moved get daily view to action
